requirements.txt
torch
torchvision
diffusers
transformers
accelerate
safetensors
fasthtml
open-clip-torch
Pillow
torch
numpy
faiss-cpu 
httpx